+++ CHANGELOG.md
@@ -23,6 +23,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - ...
 
+## [1.1.0] - 2025-06-06
+
+### Added
+- String helpers: stripHead, stripTail, quoteAll
+- Collection helpers: concatStream
+
 ## [1.0.0] - 2025-06-03
 
 ### Added